Add RepresentationInterface with toArray function
