+++ src/Sulu/Component/Rest/ListBuilder/RepresentationInterface.php
@@ -1,0 +1,20 @@
+<?php
+
+/*
+ * This file is part of Sulu.
+ *
+ * (c) Sulu GmbH
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
+namespace Sulu\Component\Rest\ListBuilder;
+
+interface RepresentationInterface
+{
+    /**
+     * @return mixed[]
+     */
+    public function toArray(): array;
+}
